Mostly serialization improvements.
Made the raycast work more closely to how it works in Unity, returning a bool instead and having the hit as an out field.
Serialization can now save and load JSON data.
A method to refresh the voxel world. Useful for loading newly modified temporary chunks.
Added a package.json file to be able to reference it through the package manager.